Reference

Base.size — Function

size(model::StateSpaceModel)

Return the dimensions n, p, m and r of the StateSpaceModel

StateSpaceModels.ztr — Function

ztr(model::StateSpaceModel)

Return the state space model arrays Z, T and R of the StateSpaceModel

StateSpaceModels.check_steady_state — Function

check_steady_state(P_t1::Matrix{T}, P_t::Matrix{T}, tol::T) where T

Return true if steady state was attained with respect to tolerance tol, false otherwise. The steady state is checked by the following equation maximum(abs.((P_t1 - P_t)./P_t1)) < tol.

StateSpaceModels.ensure_pos_sym! — Function

ensure_pos_sym!(M::Matrix{T}; ϵ::T = 1e-8) where T

Ensure that matrix M is positive and symmetric to avoid numerical errors when numbers are small by doing (M + M')/2 + ϵ*I

StateSpaceModels.prepare_forecast — Function

prepare_forecast(ss::StateSpace{Typ}, N::Int) where Typ

Adjust matrix Z for forecasting and check for dimension errors.