Manual

Manual

Introduction

In this package we consider the following state-space model

\[\begin{gather*} \begin{aligned} y_t &= Z_t \alpha_t + \varepsilon_t, \quad \quad \quad t = 1 \dots n, \\ \alpha_{t+1} &= T \alpha_t + R \eta_t, \end{aligned} \end{gather*}\]

where

\[\begin{bmatrix} \varepsilon_t \\ \eta_t \\ \alpha_1 \end{bmatrix} \sim NID \begin{pmatrix} \begin{bmatrix} 0 \\ 0 \\ a_1 \end{bmatrix} , \begin{bmatrix} H & 0 & 0\\ 0 & Q & 0\\ 0 & 0 & P_1\\ \end{bmatrix} \end{pmatrix}\]

Data structures

StateSpaceModels.StateSpaceDimensions — Type.

StateSpaceDimensions

StateSpaceModel dimensions, following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

n is the number of observations

p is the dimension of the observation vector $y_t$

m is the dimension of the state vector $\alpha_t$

r is the dimension of the state covariance matrix $Q$

source

StateSpaceModels.StateSpaceModel — Type.

StateSpaceModel

Following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

y A $n \times p$ matrix containing observations

Z A $p \times m \times n$ matrix

T A $m \times m$ matrix

R A $m \times r$ matrix

A StateSpaceModel object can be defined using StateSpaceModel(y::Matrix{Float64}, Z::Array{Float64, 3}, T::Matrix{Float64}, R::Matrix{Float64}).

Alternatively, if Z is time-invariant, it can be input as a single $p \times m$ matrix.

source

StateSpaceModels.StateSpaceCovariance — Type.

StateSpaceCovariance

Following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

H Covariance matrix of the observation vector

Q Covariance matrix of the state vector

source

StateSpaceModels.SmoothedState — Type.

SmoothedState

Following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

alpha Expected value of the smoothed state $E(\alpha_t|y_1, \dots , y_n)$

V Error covariance matrix of smoothed state $Var(\alpha_t|y_1, \dots , y_n)$

source

StateSpaceModels.FilterOutput — Type.

FilterOutput

Following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

a Predictive state $E(\alpha_t|y_{t-1}, \dots , y_1)$

att Filtered state $E(\alpha_t|y_{t}, \dots , y_1)$

v Prediction error $v_t = y_t − Z_t a_t$

P Covariance matrix of predictive state $P = Var(\alpha_t|y_{t−1}, \dots , y_1)$

Ptt Covariance matrix of filtered state $P = Var(\alpha_t|y_{t}, \dots , y_1)$

F Variance of prediction error $Var(v_{t})$

steadystate Boolean to indicate if steady state was attained

tsteady Instant when steady state was attained; in case it was not, tsteady = n+1

source

StateSpaceModels.StateSpace — Type.

StateSpace

A state-space structure containing the model, filter output, smoother output, covariance matrices, filter type and optimization method.

source

Predefined models

The local level model is defined by

\[\begin{gather*} \begin{aligned} y_t &= \mu_t + \varepsilon_t \quad \varepsilon_t \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\ \mu_{t+1} &= \mu_t + \eta_t \quad \eta_t \sim \mathcal{N}(0, \sigma^2_{\eta})\\ \end{aligned} \end{gather*}\]

StateSpaceModels.local_level — Function.

local_level(y::VecOrMat{Typ}) where Typ <: AbstractFloat

Build state-space system for a local level model with observations y.

If y is proided as an Array{Typ, 1} it will be converted to an Array{Typ, 2} inside the StateSpaceModel.

source

The linear trend model is defined by

\[\begin{gather*} \begin{aligned} y_t &= \mu_t + \varepsilon_t \quad &\varepsilon_t \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\ \mu_{t+1} &= \mu_t + \nu_t + \xi_t \quad &\xi_t \sim \mathcal{N}(0, \sigma^2_{\xi})\\ \nu_{t+1} &= \nu_t + \zeta_t \quad &\zeta_t \sim \mathcal{N}(0, \sigma^2_{\zeta})\\ \end{aligned} \end{gather*}\]

StateSpaceModels.linear_trend — Function.

linear_trend(y::VecOrMat{Typ}) where Typ <: AbstractFloat

Build state-space system for a linear trend model with observations y.

If y is proided as an Array{Typ, 1} it will be converted to an Array{Typ, 2} inside the StateSpaceModel.

source

The structural model is defined by

<!– TODO mathematical model –>

StateSpaceModels.structural — Function.

structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Float64}(undef, 0, 0)) where Typ <: AbstractFloat

Build state-space system for a given structural model with observations y, seasonality s, and, optionally, exogenous variables X.

If y is provided as an Array{Typ, 1} it will be converted to an Array{Typ, 2} inside the StateSpaceModel. The same will happen to X, if an Array{Typ, 1} it will be converted to an Array{Typ, 2} inside the StateSpaceModel.

source

Estimation

The model estimation is made using the function statespace(model; filter_type = KalmanFilter, optimization_method = RandomSeedsLBFGS(), verbose = 1). It receives as argument the pre-specified StateSpaceModel object model. Optionally, the user can define the Kalman filter variant to be used, the optimization method and the verbosity level.

StateSpaceModels.statespace — Function.

statespace(model::StateSpaceModel; filter_type::DataType = KalmanFilter, optimization_method::AbstractOptimizationMethod = RandomSeedsLBFGS(), verbose::Int = 1)

Estimate the pre-specified state-space model.

source

Forecasting

Forecasting is conducted with the function forecast. It receives as argument a StateSpace object and the number of steps ahead N.

StateSpaceModels.forecast — Function.

forecast(ss::StateSpace, N::Int)

Obtain the minimum mean square error forecasts N steps ahead. Returns the forecasts and the predictive distributions at each time period.

source

Simulation

Simulation is made using the function simulate. It receives as argument a StateSpace object, the number of steps ahead N and the number of scenarios to simulate S.

StateSpaceModels.simulate — Function.

simulate(ss::StateSpace, N::Int, S::Int)

Simulate S future scenarios up to N steps ahead. Returns a p x N x S matrix where the dimensions represent, respectively, the number of series in the model, the number of steps ahead, and the number of scenarios.

source

Filters

Kalman Filter

StateSpaceModels.kalman_filter — Function.

kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; tol::Float64 = 1e-5) where Typ <: AbstractFloat

Kalman filter with big Kappa initialization, i.e., initializing state variances as 1e6.

source

StateSpaceModels.smoother — Function.

smoother(model::StateSpaceModel, kfilter::KalmanFilter)

Smoother for state-space model.

source

Square Root Kalman Filter

StateSpaceModels.sqrt_kalman_filter — Function.

sqrt_kalman_filter(model::StateSpaceModel, sqrtH::Matrix{Typ}, sqrtQ::Matrix{Typ}; tol::Float64 = 1e-5) where Typ <: AbstractFloat

Square-root Kalman filter with big Kappa initialization.

source

StateSpaceModels.sqrt_smoother — Function.

sqrt_smoother(model::StateSpaceModel, sqrt_filter::SquareRootFilter)

Square-root smoother for state space model.

source

StateSpaceModels.filtered_state — Function.

filtered_state(model::StateSpaceModel, sqrt_filter::SquareRootFilter)

Obtain the filtered state estimates and their covariance matrices.

source

Filter interface

StateSpaceModels has an interface that allows users to define their own versions of the Kalman filter.

StateSpaceModels.AbstractFilter — Type.

AbstractFilter

Abstract type used to implement an interface for generic filters.

source

StateSpaceModels.AbstractSmoother — Type.

AbstractSmoother

Abstract type used to implement an interface for generic smoothers.

source

StateSpaceModels.kfas — Function.

kfas(model::StateSpaceModel, covariance::StateSpaceCovariance,
     filter_type::DataType)

Perform kalman filter and smoother according to the chosen filter_type.

source

Every filter must provide its own version of the statespace_likelihood function

StateSpaceModels.statespace_likelihood — Function.

statespace_likelihood(psitilde::Vector{T}, model::StateSpaceModel) where T <: AbstractFloat

Compute log-likelihood concerning hyperparameter vector psitilde ($\psi$)

Evaluate $\ell(\psi;y_n)= -\frac{np}{2}\log2\pi - \frac{1}{2} \sum_{t=1}^n \log |F_t| - \frac{1}{2} \sum_{t=1}^n v_t^{T} F_t^{-1} v_t$

source

Optimization methods

RandomSeedsLBFGS

StateSpaceModels.RandomSeedsLBFGS — Type.

TODO

source

Optimization methods interface

StateSpaceModels.AbstractOptimizationMethod — Type.

AbstractOptimizationMethod

Abstract type used to implement an interface for generic optimization methods.

source

StateSpaceModels.estimate_statespace — Function.

estimate_statespace(model::StateSpaceModel, filter_type::DataType,
                    optimization_method::AbstractOptimizationMethod; verbose::Int = 1)

Estimate parameters of the StateSpaceModel according to its filter_type and optimization_method.

source