Manual

Manual

Introduction

In this package we consider the following state space model

\[\begin{gather*} \begin{aligned} y_t &= Z_t \alpha_t + \varepsilon_t \quad \quad \quad t = 1 \dots n\\ \alpha_{t+1} &= T_t \alpha_t + R_t \eta_t \end{aligned} \end{gather*}\]

where

\[\begin{bmatrix} \varepsilon_t \\ \eta_t \\ \alpha_1 \end{bmatrix} \sim NID \begin{pmatrix} \begin{bmatrix} 0 \\ 0 \\ a_1 \end{bmatrix} , \begin{bmatrix} H_t & 0 & 0\\ 0 & Q_t & 0\\ 0 & 0 & P_1\\ \end{bmatrix} \end{pmatrix}\]

Data Structures

StateSpaceModels.StateSpaceDimensions — Type.

StateSpaceDimensions

StateSpaceModel dimensions, following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

n is the number of observations

p is the dimension of the observation vector $y_t$

m is the dimension of the state vector $\alpha_t$

r is the dimension of the state covariance matrix $Q_t$

source

StateSpaceModels.StateSpaceModel — Type.

StateSpaceModel

Following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

y A $n \times p$ matrix containing observations

Z A $p \times m \times n$ matrix

T A $m \times m$ matrix

R A $m \times r$ matrix

A StateSpaceModel object can be defined using StateSpaceModel(y::Matrix{Float64}, Z::Array{Float64, 3}, T::Matrix{Float64}, R::Matrix{Float64}).

Alternatively, if Z is time-invariant, it can be input as a single $p \times m$ matrix.

source

Missing docstring.

Missing docstring for StateSpaceParameters. Check Documenter's build log for details.

StateSpaceModels.SmoothedState — Type.

SmoothedState

Following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

alpha Expected value of the smoothed state $E(\alpha_t|y_1, \dots , y_n)$

V Error covariance matrix of smoothed state $Var(\alpha_t|y_1, \dots , y_n)$

source

StateSpaceModels.FilterOutput — Type.

Filter output

Following the notation of on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.

a predictive states $E(\alpha_t|y_{t-1}, \dots , y_1)$

v Prediction errors $v_t = y_t − Z_ta_t,i, i = 1, \dots , p$

P Covariance matrix of predictive states $P = Var(\alpha_t|y_{t−1}, \dots , y_1)$

F Prediction error variances $Var(v_t)$

steadystate Boolean to indicate if steady state was attained

tstady Instant when steady state was attained; in case it wasn't, tsteady = n+1

source

StateSpaceModels.StateSpace — Type.

StateSpace

StateSpaceModel

source

Default models

The local level model is defined by

\[\begin{gather*} \begin{aligned} y_t &= \mu_t + \varepsilon_t \quad \varepsilon_t \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\ \mu_{t+1} &= \mu_t + \eta_t \quad \eta_t \sim \mathcal{N}(0, \sigma^2_{\eta})\\ \end{aligned} \end{gather*}\]

StateSpaceModels.local_level — Function.

local_level(y::VecOrMat{Typ}) where Typ <: AbstractFloat

Build state-space system for a local level model with observations y.

If y is proided as an Array{Typ, 1} it will be converted to an Array{Typ, 2} inside the StateSpaceModel.

source

The linear trend model is defined by

\[\begin{gather*} \begin{aligned} y_t &= \mu_t + \varepsilon_t \quad &\varepsilon_t \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\ \mu_{t+1} &= \mu_t + \nu_t + \xi_t \quad &\xi_t \sim \mathcal{N}(0, \sigma^2_{\xi})\\ \nu_{t+1} &= \nu_t + \zeta_t \quad &\zeta_t \sim \mathcal{N}(0, \sigma^2_{\zeta})\\ \end{aligned} \end{gather*}\]

StateSpaceModels.linear_trend — Function.

linear_trend(y::VecOrMat{Typ}) where Typ <: AbstractFloat

Build state-space system for a linear trend model with observations y.

If y is proided as an Array{Typ, 1} it will be converted to an Array{Typ, 2} inside the StateSpaceModel.

source

The structural model is defined by

<!– TODO mathematical model –>

Missing docstring.

Missing docstring for structuralmodel. Check Documenter's build log for details.

Estimation

The model estimation is made using the function statespace(y, s; X, nseeds). It receives as argument the time series y and the desired seasonality s. The user can input exogenous variables using optional argument X and specify the desired number of random seeds nseeds to perform the estimation.

<!– TODO –>

Simulation

Simulation is made using the function simulate. It receives as argument a StateSpace object, the number of steps ahead N and the number of scenarios to simulate S.

simulation = simulate(ss, N, S)

Filters

<!– TODO sqrt kalman filter, put recursion and reference –> <!– TODO bigkappa kalman filter, put recursion and reference –>

Optimization methods

<!– LBFGS put reference and Optim manual –>