Reference

Reference

<!– utils.jl –>

Base.size — Function.

size(model::StateSpaceModel)

Return the dimensions n, p, m and r of the StateSpaceModel

source

StateSpaceModels.ztr — Function.

ztr(model::StateSpaceModel)

Return the state space model arrays Z, T and R of the StateSpaceModel

source

Missing docstring.

Missing docstring for check_steady_state. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ensure_pos_sym. Check Documenter's build log for details.